backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
httpx==0.27.2
python-dotenv==1.0.1
boto3==1.42.96
requests==2.31.0
bcrypt==4.1.2
PyJWT==2.8.0
cryptography==44.0.0
Pillow==10.2.0
email-validator>=2.0.0

